Update README.md
Nowadays, there is a lot of open data about agriculture statistics in the world. However, few tools are presented to predict and visualize these processes. This project will show how to download data from open sources, perform preliminary data analysis, transform and clear data, perform correlation. Next we will consider 2 different mathematical approaches to the calculation of a forecast based on linear regression. To do this, the division of the DataSet into training and test sets will be demonstrated. It will be shown how to build models using 2 different frameworks. Then we will build a forecast and analyze the accuracy and adequacy of the obtained models. Then two different Neuro Networks will be build and fitted. We will study how to normalize and inverse normalize data. At the end we will compare all the results.